SONAR-5867 SONAR-5819 Index symbol highlighting and exploit it in api/sources/lines

+++ server/sonar-server/src/test/java/org/sonar/server/source/SourceServiceTest.java
@@ -89,7 +89,7 @@ public class SourceServiceTest {
 
     service.getLinesAsHtml(COMPONENT_KEY);
 
-    verify(sourceDecorator).getDecoratedSourceAsHtml(COMPONENT_KEY, null, null);
+    verify(sourceDecorator).getDecoratedSourceAsHtml(COMPONENT_KEY, (Integer) null, (Integer) null);
   }
 
   @Test
